Crate aws_sdk_applicationautoscaling

With Application Auto Scaling, you can configure automatic scaling for the following resources:

Amazon ECS services

Amazon EC2 Spot Fleet requests

Amazon EMR clusters

Amazon AppStream 2.0 fleets

Amazon DynamoDB tables and global secondary indexes throughput capacity

Amazon Aurora Replicas

Amazon SageMaker endpoint variants

Custom resources provided by your own applications or services

Amazon Comprehend document classification and entity recognizer endpoints

AWS Lambda function provisioned concurrency

Amazon Keyspaces (for Apache Cassandra) tables

Amazon Managed Streaming for Apache Kafka broker storage

API Summary

The Application Auto Scaling service API includes three key sets of actions:

Register and manage scalable targets - Register AWS or custom resources as scalable targets (a resource that Application Auto Scaling can scale), set minimum and maximum capacity limits, and retrieve information on existing scalable targets.

Configure and manage automatic scaling - Define scaling policies to dynamically scale your resources in response to CloudWatch alarms, schedule one-time or recurring scaling actions, and retrieve your recent scaling activity history.

Suspend and resume scaling - Temporarily suspend and later resume automatic scaling by calling the RegisterScalableTarget API action for any Application Auto Scaling scalable target. You can suspend and resume (individually or in combination) scale-out activities that are triggered by a scaling policy, scale-in activities that are triggered by a scaling policy, and scheduled scaling.

To learn more about Application Auto Scaling, including information about granting IAM users required permissions for Application Auto Scaling actions, see the Application Auto Scaling User Guide.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

config

error

input

model

operation

output

Structs

Blob

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

SdkError

Failing Sdk Result

Statics

PKG_VERSION