Module aws_sdk_applicationautoscaling::error

Modules

concurrent_update_exception

See ConcurrentUpdateException

failed_resource_access_exception

See FailedResourceAccessException

internal_service_exception

See InternalServiceException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

object_not_found_exception

See ObjectNotFoundException

validation_exception

See ValidationException

Structs

ConcurrentUpdateException

Concurrent updates caused an exception, for example, if you request an update to an

DeleteScalingPolicyError

DeleteScheduledActionError

DeregisterScalableTargetError

DescribeScalableTargetsError

DescribeScalingActivitiesError

DescribeScalingPoliciesError

DescribeScheduledActionsError

FailedResourceAccessException

Failed access to resources caused an exception. This exception is thrown when Application Auto Scaling

InternalServiceException

The service encountered an internal error.

InvalidNextTokenException

The next token supplied was invalid.

LimitExceededException

A per-account resource limit is exceeded. For more information, see Application Auto Scaling service quotas.

ObjectNotFoundException

The specified object could not be found. For any operation that depends on the existence

PutScalingPolicyError

PutScheduledActionError

RegisterScalableTargetError

ValidationException

An exception was thrown for a validation issue. Review the available parameters for the

Enums

DeleteScalingPolicyErrorKind

DeleteScheduledActionErrorKind

DeregisterScalableTargetErrorKind

DescribeScalableTargetsErrorKind

DescribeScalingActivitiesErrorKind

DescribeScalingPoliciesErrorKind

DescribeScheduledActionsErrorKind

PutScalingPolicyErrorKind

PutScheduledActionErrorKind

RegisterScalableTargetErrorKind