Module aws_sdk_applicationautoscaling::model

Modules

alarm

See Alarm

customized_metric_specification

See CustomizedMetricSpecification

metric_dimension

See MetricDimension

predefined_metric_specification

See PredefinedMetricSpecification

scalable_target

See ScalableTarget

scalable_target_action

See ScalableTargetAction

scaling_activity

See ScalingActivity

scaling_policy

See ScalingPolicy

scheduled_action

See ScheduledAction

step_adjustment

See StepAdjustment

step_scaling_policy_configuration

See StepScalingPolicyConfiguration

suspended_state

See SuspendedState

target_tracking_scaling_policy_configuration

See TargetTrackingScalingPolicyConfiguration

Structs

Alarm

Represents a CloudWatch alarm associated with a scaling policy.

CustomizedMetricSpecification

Represents a CloudWatch metric of your choosing for a target tracking scaling policy to use

MetricDimension

Describes the dimension names and values associated with a metric.

PredefinedMetricSpecification

Represents a predefined metric for a target tracking scaling policy to use with

ScalableTarget

Represents a scalable target.

ScalableTargetAction

Represents the minimum and maximum capacity for a scheduled action.

ScalingActivity

Represents a scaling activity.

ScalingPolicy

Represents a scaling policy to use with Application Auto Scaling.

ScheduledAction

Represents a scheduled action.

StepAdjustment

Represents a step adjustment for a StepScalingPolicyConfiguration. Describes an adjustment based on the difference

StepScalingPolicyConfiguration

Represents a step scaling policy configuration to use with Application Auto Scaling.

SuspendedState

Specifies whether the scaling activities for a scalable target are in a suspended state.

TargetTrackingScalingPolicyConfiguration

Represents a target tracking scaling policy configuration to use with Application Auto Scaling.

Enums

AdjustmentType

MetricAggregationType

MetricStatistic

MetricType

PolicyType

ScalableDimension

ScalingActivityStatusCode

ServiceNamespace