Module aws_sdk_applicationautoscaling::operation

Structs

DeleteScalingPolicy

Deletes the specified scaling policy for an Application Auto Scaling scalable target.

DeleteScheduledAction

Deletes the specified scheduled action for an Application Auto Scaling scalable target.

DeregisterScalableTarget

Deregisters an Application Auto Scaling scalable target when you have finished using it. To see which

DescribeScalableTargets

Gets information about the scalable targets in the specified namespace.

DescribeScalingActivities

Provides descriptive information about the scaling activities in the specified namespace

DescribeScalingPolicies

Describes the Application Auto Scaling scaling policies for the specified service namespace.

DescribeScheduledActions

Describes the Application Auto Scaling scheduled actions for the specified service namespace.

PutScalingPolicy

Creates or updates a scaling policy for an Application Auto Scaling scalable target.

PutScheduledAction

Creates or updates a scheduled action for an Application Auto Scaling scalable target.

RegisterScalableTarget

Registers or updates a scalable target.